pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "exec-env"
version = "0.1.0"
description = "ExecEnv: AI Executive Assistant Benchmark"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "openenv-core",
    "pydantic>=2.0.0",
    "openai",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "websockets>=11.0.0",
    "httpx>=0.24.0",
    "python-dotenv"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
py-modules = ["exec_env", "models", "tasks", "inference"]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
